gui/pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "scapy-studio-gui"
version = "0.1.0"
description = "Scapy 的图形化工作台骨架"
readme = "README.rst"
requires-python = ">=3.9"
dependencies = [
    "PySide6-Essentials>=6.7,<7",
    "scapy",
]

[project.scripts]
packet-studio = "packet_studio.__main__:main"
packet-studio-validation = "packet_studio.validation_launcher:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
include = ["packet_studio*"]